# storage/api-docs-v1/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Portal API V1 Swagger Documentation",
        "version": "1"
    },
    "paths": {
...
        "/api/v1/comments": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Get list comments ",
                "description": "Endpoint to get list comments ",
                "operationId": "31364881058187fc38a42287e9d904a1",
                "parameters": [
                    {
                        "name": "article_id",
                        "in": "query",
                        "description": "Article ID"
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "Content"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "1 for Ascending -1 for Descending"
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit (Default 10)"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Num Of Page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get list comments successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        },
                                        "pagination": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Get list comments failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Get list comments failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Create a new comment",
                "description": "Create a new comment",
                "operationId": "51f741718cabbf7ddf0e49b69f645dbe",
                "requestBody": {
                    "description": "Data that needed to create a new comment",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "article_id",
                                    "content"
                                ],
                                "properties": {
                                    "article_id": {
                                        "type": "number",
                                        "example": 1
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "comment"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create a new comment successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Create a new comment successfully"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Create a new comment failed"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/v1/comments/{id}/reply": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Reply to comment",
                "description": "Reply to comment",
                "operationId": "feb6cc77e5314a484840773a4e0ab517",
                "requestBody": {
                    "description": "Data that needed to create a new reply",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "parent_id",
                                    "content"
                                ],
                                "properties": {
                                    "parent_id": {
                                        "type": "number",
                                        "example": 1
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "comment"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create a new reply successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Create a new reply successfully"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Create a new reply failed"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/v1/comments/{id}/toggle-reaction": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Add reaction to comment",
                "description": "Endpoint to add reaction to comment",
                "operationId": "a0310d48acd440b866c62b9cc85d42f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID"
                    }
                ],
                "requestBody": {
                    "description": "Data that needed to add reaction",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "like"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add reaction successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Add reaction successfully"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Add reaction failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Add reaction failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}